+++ resources/3rdparty/sylvan/src/storm_wrapper.cpp
@@ -215,8 +215,8 @@ storm_rational_number_ptr storm_rational_number_pow(storm_rational_number_ptr a,
     storm::RationalNumber const& srn_a = *(storm::RationalNumber const*)a;
     storm::RationalNumber const& srn_b = *(storm::RationalNumber const*)b;
 
-    carl::uint exponentAsInteger = carl::toInt<carl::uint>(srn_b);
-    storm::RationalNumber* result_srn = new storm::RationalNumber(carl::pow(srn_a, exponentAsInteger));
+    carl::sint exponentAsInteger = carl::toInt<carl::sint>(srn_b);
+    storm::RationalNumber* result_srn = new storm::RationalNumber(storm::utility::pow(srn_a, exponentAsInteger));
     return (storm_rational_number_ptr)result_srn;
 }